+++ src/MessageFlags.lua
@@ -1,0 +1,17 @@
+--!strict
+
+return {
+	Crossposted = 1,
+	IsCrosspost = 2,
+	SuppressEmbeds = 4,
+	SourceMessageDeleted = 8,
+	Urgent = 16,
+	HasThread = 32,
+	Ephemeral = 64,
+	Loading = 128,
+	FailedToMentionSomeRolesInThread = 256,
+	SuppressNotifications = 4096,
+	IsVoiceMessage = 8192,
+	HasSnapshot = 16384,
+	IsComponentsV2 = 32768,
+}